+++ src/Symfony/Component/Form/Extension/PasswordHasher/Type/FormTypePasswordHasherExtension.php
@@ -19,16 +19,23 @@ use Symfony\Component\Form\FormEvents;
 
 /**
  * @author Sébastien Alfaiate <s.alfaiate@webarea.fr>
+ *
+ * @deprecated since Symfony 8.2, passwords are hashed during the "form.post_validate" event, which requires the validator extension
  */
 class FormTypePasswordHasherExtension extends AbstractTypeExtension
 {
     public function __construct(
         private PasswordHasherListener $passwordHasherListener,
     ) {
+        trigger_deprecation('symfony/form', '8.2', 'The "%s" class is deprecated, passwords are hashed during the "form.post_validate" event instead.', self::class);
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
+        if ($options['hash_property_path'] ?? false) {
+            $builder->addEventListener(FormEvents::POST_SUBMIT, [$this->passwordHasherListener, 'registerPassword']);
+        }
+
         $builder->addEventListener(FormEvents::POST_SUBMIT, [$this->passwordHasherListener, 'hashPasswords']);
     }