+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG
  * Add the `form_id` view variable, holding the id to render on the `<form>` element of a root form when a child uses `form_attr`
  * Allow the `group_by` option of `ChoiceType` to return `TranslatableInterface` instances
  * Add the `FormEvents::POST_VALIDATE` event, dispatched on each form of the tree after the validation of the root form
+ * Deprecate the `FormTypePasswordHasherExtension` class and the `registerPassword()` and `hashPasswords()` methods of `PasswordHasherListener`, passwords are now hashed during the `form.post_validate` event
 
 8.1
 ---